+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [1.1.0]
 * Updated `Sign in with Instagram` screen
-* Added pagination support (up to 20 more photos each load)  
+* Added pagination support (up to 20 more photos each load)
+* Updated selection UI  
 
 ## [1.0.0]
 * Using __new REST API solution__ to fetch the photos correctly